requirements.txt
customtkinter
matplotlib
pyyaml
pandas